% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_font.R
\name{get_standard_ja_fonts}
\alias{get_standard_ja_fonts}
\title{get OS standard font families}
\usage{
get_standard_ja_fonts()
}
\value{
character vector of length 2. These elements are names of (1) Sans (Gothic) and (2) Serif (Mincho), Japanese font families which are installed standardly in your operating system.
}
\description{
get OS-tandard Japanese font family names
}
\details{
This function returns character vector of length 2 which are standard Japanese font families for each operating system. Thus these families are OS-specified:\tabular{lcc}{
   OS \tab Sans \tab Serif \cr
   Linux \tab \code{"Noto Sans CJK JP"} \tab \code{"Noto Serif CJK JP"} \cr
   Mac \tab \code{"Hiragino Sans"} \tab \code{"Hiragino Mincho ProN"} \cr
   Windows (>=8) \tab \code{"Yu Gothic"} \tab \code{"Yu Gothic"} \cr
   Windows (<8) \tab \code{"MS Gothic"} \tab \code{"MS Mincho"} \cr
}
}
