% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{get_styles}
\alias{get_styles}
\title{get font names corresponding to four style names for a font family/フォントファミリから4種類の書体に対応するフォント名を取り出す}
\usage{
get_styles(data)
}
\arguments{
\item{data:}{`data.frame` alike, output of font list from `systemfonts::system_fonts()` or its subset, which includes six columns: `name`, `family`, `style`, `weight`, `width`, and `italic`/ `data.frame` 相当のオブジェクトで, `systemfonts::system_fonts()` のから特定のファミリだけ取り出したもので, `name`, `family`, `style`, `weight`, `width`, `italic` 列を持つもの.}
}
\value{
character vector of length 4: It's elements are names of plain, bold, italic, and bold italic font./長さ4の文字列ベクトル. 要素は順に標準, ボールド, イタリック, ボールドイタリック体のpostscript名.
}
\description{
`systemfonts::system_fonts()` などの結果から4書体のフォントを取得する. ただしうまくいくとは限らない.
}
