% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_all_fonts}
\alias{register_all_fonts}
\title{Register Fonts / フォントを登録する}
\usage{
register_all_fonts()
}
\description{
register all system fonts obtained from `systemfonts` package by `windowsFonts` or `quartzFonts`/ `systemfonts` で取得したフォント名を全てOS固有のデバイス (`windowsFonts`, `quartzFonts`) に登録する.
}
