% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_font.R
\name{set_standard_ja_fonts}
\alias{set_standard_ja_fonts}
\title{set standard \code{sans}/\code{serif} family so that to refer the OS standard font}
\usage{
set_standard_ja_fonts(sans = NULL, serif = NULL)
}
\description{
change default Serif/Sans Serif fonts for Windows & Mac OS / Windows および Mac で \code{sans}/\code{serif} の参照するフォントを変更する. デフォルトはOSごとの標準日本語フォント
}
